Time tracking is improved for threads
The old approach was tracking the sleeping time for
other threads while one is working, which is an intersection of time
tracking that leads into counting more the time of the computation which
in reality took less time.
In the new approach the time tracking is more a representative the
amount time the computation took.